# docs/api.yaml
swagger: "2.0"
info:
  description: "Aplikasi ini digunakan untuk menyimpan foto maupun membuat comment untuk foto orang lain."
  version: "1.0"
  title: "MyGram"
  contact:
    email: "abdul.rasyid.anshori@gmail.com"
host: "localhost"
basePath: "/api/v1.0"
...
schemes:
- "https"
- "http"
paths:
...
  /users:
    put:
      tags:
      - "user"
      summary: "Update User"
      description: "Endpoint ini digunakan untuk update data user."
      operationId: "updateUser"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Bearer"
          type: "string"
          description: "Masukan Bearer Token yang diterima pada saat login"
          required: true
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/UserUpdate"
          required: true
      responses:
        "200":
          description: "OK"
          schema: 
            $ref: "#/definitions/UserUpdateResponse"
      security:
      - Bearer: []
    delete:
      tags:
      - "user"
      summary: "Delete User"
      description: "Endpoint ini digunakan untuk delete user."
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Bearer"
          type: "string"
          description: "Masukan Bearer Token yang diterima pada saat login"
          required: true
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Your account has been successfully deleted"
      security:
      - Bearer: []
...
  /photos/{photoId}:
    put:
      tags:
      - "photo"
      summary: "Update Photo"
      description: "Endpoint ini digunakan untuk update photo."
      operationId: "updatePhoto"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Bearer"
          type: "string"
          description: "Masukan Bearer Token yang diterima pada saat login"
          required: true
        - in: "path"
          name: "photoId"
          description: "ID photo yang akan di update"
          required: true
          type: "integer"
          format: "uint"
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/PhotoCreateAndUpdate"
      responses:
        "200":
          description: "OK"
          schema: 
            $ref: "#/definitions/PhotoUpdateResponse"
      security:
      - Bearer: []
    delete:
      tags:
      - "photo"
      summary: "Delete Photo"
      description: "Endpoint ini digunakan untuk menghapus photo."
      operationId: "deletePhoto"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Bearer"
          type: "string"
          description: "Masukan Bearer Token yang diterima pada saat login"
          required: true
        - in: "path"
          name: "photoId"
          description: "ID photo yang akan di delete"
          required: true
          type: "integer"
          format: "uint"
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Your photo has been successfully deleted"
      security:
      - Bearer: []
...
  /comments/{commentId}:
    put:
      tags:
      - "comment"
      summary: "Update Photo"
      description: "Endpoint ini digunakan untuk update comment."
      operationId: "updatePhoto"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Bearer"
          type: "string"
          description: "Masukan Bearer Token yang diterima pada saat login"
          required: true
        - in: "path"
          name: "commentId"
          description: "ID comment yang akan di update"
          required: true
          type: "integer"
          format: "uint"
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/CommentUpdate"
      responses:
        "200":
          description: "OK"
          schema: 
            $ref: "#/definitions/CommentUpdateResponse"
      security:
      - Bearer: []
    delete:
      tags:
      - "comment"
      summary: "Delete Photo"
      description: "Endpoint ini digunakan untuk menghapus comment."
      operationId: "deletePhoto"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Bearer"
          type: "string"
          description: "Masukan Bearer Token yang diterima pada saat login"
          required: true
        - in: "path"
          name: "commentId"
          description: "ID comment yang akan di delete"
          required: true
          type: "integer"
          format: "uint"
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Your comment has been successfully deleted"
      security:
      - Bearer: []
...
  /socialmedias/{socialMediaId}:
    put:
      tags:
      - "social media"
      summary: "Update Social Media"
      description: "Endpoint ini digunakan untuk update social media."
      operationId: "updateSocialMedia"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Bearer"
          type: "string"
          description: "Masukan Bearer Token yang diterima pada saat login"
          required: true
        - in: "path"
          name: "socialMediaId"
          description: "ID social media yang akan di update"
          required: true
          type: "integer"
          format: "uint"
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/SocialMediaCreateAndUpdate"
      responses:
        "200":
          description: "OK"
          schema: 
            $ref: "#/definitions/SocialMediaUpdateResponse"
      security:
      - Bearer: []
    delete:
      tags:
      - "social media"
      summary: "Delete Social Media"
      description: "Endpoint ini digunakan untuk menghapus social media."
      operationId: "deleteSocialMedia"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Bearer"
          type: "string"
          description: "Masukan Bearer Token yang diterima pada saat login"
          required: true
        - in: "path"
          name: "socialMediaId"
          description: "ID social media yang akan di delete"
          required: true
          type: "integer"
          format: "uint"
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Your social media has been successfully deleted"
      security:
      - Bearer: []
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  UserUpdate:
    type: "object"
    properties:
      email:
        type: "string"
      username:
        type: "string"
  UserUpdateResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      email:
        type: "string"
      username:
        type: "string"
      age:
        type: "integer"
      updated_at:
        type: "string"
        example: "date"
...
  PhotoCreateAndUpdate:
    type: "object"
    properties:
      title:
        type: "string"
      caption:
        type: "string"
      photo_url:
        type: "string"
...
  PhotoUpdateResponse:
    allOf:
    - type: object
      properties:
        id:
          type: "integer"
    - $ref: "#/definitions/PhotoCreateAndUpdate"
    - type: object
      properties:
        user_id:
          type: "integer"
        updated_at:
          type: "string"
          example: "Date"
  CommentCreate:
    type: "object"
    properties:
      message:
        type: "string"
      photo_id:
        type: "integer"
...
  CommentUpdate:
    type: "object"
    properties:
      message:
        type: "string"
  CommentUpdateResponse:
    allOf:
    - type: object
      properties:
        id:
          type: "integer"
    - $ref: "#/definitions/CommentCreate"
    - type: object
      properties:
        user_id:
          type: "integer"
        updated_at:
          type: "string"
          example: "Date"
...
  SocialMediaCreateAndUpdate:
    type: "object"
    properties:
      name:
        type: "string"
      social_media_url:
        type: "string"
...
  SocialMediaUpdateResponse:
    allOf:
    - type: object
      properties:
        id:
          type: "integer"
    - $ref: "#/definitions/SocialMediaCreateAndUpdate"
    - type: object
      properties:
        user_id:
          type: "integer"
        updated_at:
          type: "string"
          example: "Date"
...